Saud

+++ src/main/java/org/example/tahadaw/DTO/IN/GiftHistoryUpdateDTOIn.java
@@ -14,7 +14,7 @@ public class GiftHistoryUpdateDTOIn {
     private String giftName;
     private String occasionType;
     private LocalDate giftDate;
-    private Long priceMinor;
+    private Double priceMinor;
     private Boolean wasGifted;
     private Integer userRating;
     private String notes;
